requirements.txt
# Use PyTorch CPU wheels (cp313) on Streamlit Cloud
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.6.0+cpu
torchvision==0.21.0+cpu
Pillow==10.4.0
streamlit==1.52.2
